+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Allow resource limits/requests to be passed as values.
+- Switch `gp2` to `gp3` volumes.
+- Allow etcd volume IOPS and Throughput to be set.
 
 ## [10.18.0] - 2022-03-04